Chore #2028: radically clean up core-macro
