--- packages/core-macro/src/component_body_deserializers/mod.rs
@@ -1,4 +1,0 @@
-//! This module contains all [`ComponentBody`](crate::component_body::ComponentBody) deserializers.
-
-pub mod component;
-pub mod inline_props;
